refactor: retire legacy provider home screen
